+++ src/UI/TransactionsToastList/components/TransactionToast/hooks/useTransactionToast.ts
@@ -6,7 +6,6 @@ import { shardSelector } from 'reduxStore/selectors';
 import { TransactionBatchStatusesEnum } from 'types';
 import { getUnixTimestamp } from 'utils/dateTime/getUnixTimestamp';
 import { getUnixTimestampWithAddedMilliseconds } from 'utils/dateTime/getUnixTimestampWithAddedMilliseconds';
-import { MAX_REST } from 'utils/math';
 import { getAreTransactionsOnSameShard } from 'utils/transactions/getAreTransactionsOnSameShard';
 import {
   getIsTransactionPending,
@@ -52,10 +51,7 @@ export const useTransactionToast = ({
 
   const progress = {
     startTime,
-    endTime,
-    animationMax: areSameShardTransactions
-      ? MAX_REST
-      : CROSS_SHARD_ROUNDS * MAX_REST
+    endTime
   };
 
   const isPending = getIsTransactionPending(status);
@@ -90,6 +86,7 @@ export const useTransactionToast = ({
   }, [lifetimeAfterSuccess, status]);
 
   return {
+    isCrossShard: !areSameShardTransactions,
     progress,
     isPending,
     isTimedOut,